scalaparse: allow whitespace after minus sign for numbers
In Scala, `- 1` is a valid literal, but the current `Literal`
expression does not allow for the whitespace after the minus sign.
This has not been a problem for expressions, since the `ExprPrefix`
rule allows whitespace following `+`, `-`, `!`, `~`. However,
with the introduction of literal types, the following valid
Scala 2.13 code does not parse:

```
val a: - 1 = -1
```

since the `Literal` rule, which is invoked through `Type`,
won't allow whitespace after the minus sign.