Make Whitespace a proper trait rather than function to avoid unwanted implicit conversions (#272)
Fixes https://github.com/com-lihaoyi/fastparse/issues/261, which is
caused by the whitespace `P[_] => P[Unit]` implicits firing as implicit
conversions rather than the implicit parameters they were originally
intended to be. Since whitespace is meant to ignore failures, this
caused failures caused by the `Fail.! : P[String]` to be ignored when
implicitly converted to `P[Unit]`

The fix is to create a proper `Whitespace` trait to inherit from. 

Note that this is a binary incompatible change. It's stacked on top of
https://github.com/com-lihaoyi/fastparse/pull/271 for convenience, but
should probably land separately after. The relevant changes are in this
commit
https://github.com/com-lihaoyi/fastparse/pull/272/commits/d87ab6d9de062ec1e0bbf897643cefef7cd03338
if anyone wants to look.

Note that I moved the custom whitespace tests to a `scala2.12+` folder
to skip them on 2.11. 2.11 does not support SAM conversion, which makes
defining custom whitespaces a lot more boilerplatey. Can still be done,
but no need to burden everyone with boilerplatey examples just to cater
for the 2.11 folks in the test suite. Things are unlikely to break just
for 2.11 anyway