Add Scala 3 to crossNativeVersions (#288)
Currently, Fastparse cannot be used to target Scala Native if one is
using Scala 3. All the native tests are passing, so it seems like it's
okay to have it build for Scala 3 too. Apologies in advance if there's
some reason for only building with Scala 2 when targeting native, I
would've asked in the Gitter room but had some trouble joining.